Move contents of finesse.gui.hardware_set.hardware_set to parent namespace
