+++ finesse/gui/hardware_set/hardware_sets_combo_box.py
@@ -3,7 +3,7 @@
 from pubsub import pub
 from PySide6.QtWidgets import QComboBox
 
-from finesse.gui.hardware_set.hardware_set import (
+from finesse.gui.hardware_set import (
     HardwareSet,
     OpenDeviceArgs,
     get_hardware_sets,
